+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 ### Added
 - The `hasChangedBelongsTo` function, in `Projection.Model`, to check 'belongsTo' relationship changes.
 
+### Changed
+- Getting of current user name is now able by calling `getCurrentUserName` method from `user` service.
+
 ### Fixed
 - Tracking changes for `syncUp` when `belongsTo` relationships changes.